Separate icepak configuration files test
